Pull docker images last in bosco morning command
- separate bosco pull into pull-git and pull-docker
- compose pull-docker and pull-git in pull so functionality remains unchanged
- use pull-docker and pull-git in bosco morning so that pulling docker images can be done last
